refactor(containers): move all generic containers in containers/
