+++ src/analyzer/p/p_self.c
@@ -5,7 +5,7 @@
 #define LOG_TAG "p"
 #include "../../logger.h"
 
-#include "../../list.h"
+#include "../../containers/list.h"
 #include "../../protocol.h"
 
 #include <assert.h>
