+++ src/analyzer/q/q_self.c
@@ -3,7 +3,7 @@
 #define LOG_TAG "q"
 #include "../../logger.h"
 
-#include "../../list.h"
+#include "../../containers/list.h"
 #include "../../protocol.h"
 
 #include <assert.h>
