+++ src/containers/CMakeLists.txt
@@ -1,0 +1,5 @@
+add_library(containers STATIC "")
+target_sources(containers
+  PRIVATE
+  list.c
+  )
